pyproject.toml
[tool.poetry]
name = "simple-chat-app-01"
version = "0.1.0"
description = ""
authors = ["nibura2002 <nibura2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
openai = "^1.30.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
